DB data shown on the dashboard in Tiebe's UI, but adapted.
