--- app/jens/dashboard/layout.tsx
@@ -1,21 +1,0 @@
-import React from "react";
-import NavigationSidebar from "@/components/navigation/navigation-sidebar";
-import NavigationSidebarAuthentication from "@/components/authentication/navigation-sidebar-authentication";
-import NavigationSidebarLink from "@/components/navigation/navigation-sidebar-link";
-
-export const dynamic = 'force-dynamic'
-
-interface LayoutProps {
-    children: React.ReactNode;
-}
-
-export default async function layout({children}: LayoutProps) {
-    return <div className={"flex flex-col sm:flex-row"}>
-        <NavigationSidebar authNode={<NavigationSidebarAuthentication/>}>
-            <NavigationSidebarLink href={"/dashboard/zalen"}>Zalen</NavigationSidebarLink>
-        </NavigationSidebar>
-        <div className={"container max-w-screen-md mx-auto p-2"}>
-            {children}
-        </div>
-    </div>
-}
