--- app/jens/dashboard/zalen/page.tsx
@@ -1,4 +1,0 @@
-export default async function page() {
-    return <div>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>Hallo<br/>
-    </div>
-}
