+++ app/tiebe/dashboard/layout.tsx
@@ -1,0 +1,16 @@
+"use client"
+
+import * as React from "react"
+
+import DashboardNavigation from "@/components/ui/dashboard/Navigation";
+
+
+export default function PublicNavigationLayoutKlant({children}: { children: React.ReactNode }) {
+
+    return <>
+        <DashboardNavigation/>
+        <div className={"sm:ml-80 h-full sm:overflow-y-auto"}>
+            {children}
+        </div>
+    </>;
+}
