+++ components/authentication/navigation-sidebar-authentication.tsx
@@ -3,8 +3,6 @@ import {cookies} from "next/headers";
 import LogoutButton from "@/components/authentication/LogoutButton";
 import NavigationSidebarLink from "@/components/navigation/navigation-sidebar-link";
 
-export const dynamic = 'force-dynamic'
-
 export default async function NavigationSidebarAuthentication() {
     const supabase = createServerComponentClient({cookies})
     const {data: {user}} = await supabase.auth.getUser()
