+++ lib/database.types.ts
@@ -1,3 +1,5 @@
+import {PostgrestError} from '@supabase/supabase-js'
+
 export type Json =
     | string
     | number
@@ -70,3 +72,12 @@ export interface Database {
     }
   }
 }
+
+export type Tables<T extends keyof Database['public']['Tables']> = Database['public']['Tables'][T]['Row']
+export type Enums<T extends keyof Database['public']['Enums']> = Database['public']['Enums'][T]
+// etc.
+
+
+export type DbResult<T> = T extends PromiseLike<infer U> ? U : never
+export type DbResultOk<T> = T extends PromiseLike<{ data: infer U }> ? Exclude<U, null> : never
+export type DbResultErr = PostgrestError
